aidan fix pt 2
